etl/ETL/requirements.txt
elasticsearch==8.10.1
SQLAlchemy==2.0.9
redis==4.3.4
python-dotenv==0.17.0
psycopg2-binary==2.9.8
requests